+++ src/resolvers.ts
@@ -11,7 +11,7 @@ type UserInput = {
   name: string;
   email: string;
   password: string;
-  date_of_birth: string;
+  dateOfBirth: string;
   profession: string;
 };
 
@@ -55,12 +55,12 @@ export const resolvers = {
       newUser.name = requestData.name;
       newUser.email = requestData.email;
       newUser.password = hash;
-      newUser.date_of_birth = new Date(requestData.date_of_birth);
+      newUser.dateOfBirth = new Date(requestData.dateOfBirth);
       newUser.profession = requestData.profession;
       const savedUser = await UserRepository.save(newUser);
       return {
         ...savedUser,
-        date_of_birth: savedUser.date_of_birth.toISOString(),
+        dateOfBirth: savedUser.dateOfBirth.toISOString(),
       };
     },
 
@@ -74,23 +74,25 @@ export const resolvers = {
       if (!existingUser) {
         throw new Error("Email não encontrado!");
       }
+      console.log(requestCredentials.password);
       const hash = createHash("sha256")
         .update(requestCredentials.password)
         .digest("hex");
+      console.log(hash);
+      console.log(existingUser.password);
       if (existingUser.password != hash) {
         throw new Error("Senha inválida!");
       }
-      const user_date_of_birth = new Date(existingUser.date_of_birth);
+      const user_dateOfBirth = new Date(existingUser.dateOfBirth);
       const data = {
-        login: {
-          user: {
-            profession: existingUser.profession,
-            name: existingUser.name,
-            email: existingUser.email,
-            date_of_birth: user_date_of_birth,
-          },
-          token: "the_token",
+        user: {
+          profession: existingUser.profession,
+          name: existingUser.name,
+          email: existingUser.email,
+          dateOfBirth: user_dateOfBirth,
+          id: existingUser.id,
         },
+        token: "the_token",
       };
       return data;
     },
