+++ src/schema.ts
@@ -3,7 +3,7 @@ input UserInput {
 	name: String!
 	email: String!
 	password: String!
-	date_of_birth: String!
+	dateOfBirth: String!
 	profession: String
 }
 
@@ -15,7 +15,7 @@ input LoginInfo {
 type User {
 	name: String
 	email: String
-	date_of_birth: String
+	dateOfBirth: String
 	profession: String
 	id: Int
 }
@@ -24,9 +24,18 @@ type Query {
 	users: String
 }
 
+type LoginData {
+	user: User
+	token: String
+}
+
 type Mutation {
 	insertUser( 
 		requestData: UserInput
 	): User
+
+	login(
+		requestCredentials: LoginInfo
+	): LoginData
 }
 `;
